Add parameters to errors so they can be handled in user code
Previously, the only way to differentiate between different possible errors inside `plex-api` was to parse the actualy `message` string. Now the `name` parameter gets set in all Errors that are returned, and in the case of HTTP status code errors, the statusCode is also passed along in the error object.